hide reporting usage info for developers
It's useful to have this info in here, but it might just serve to confuse developers. For this reason I've hidden it using an [admonition](https://squidfunk.github.io/mkdocs-material/reference/admonitions/). FWIW I think it would be cool to extend this pattern to other docs pages to clue them into how these events inform reporting.